feat(0.1.0): release candidate for Router
This change pulls the core ideas of pick into a single router.ts file
and implements what I think is a decent interface for the original
Router. In the process I moved the indexed async state either monad, now
called Effect, to the fun library. Effect is now variadic in a clever
but very useful way. Many of the interfaces for Handler, Route, and
Router have breaking changes. I've implemented a very basic but quite
performant default router that does fallthrough routing after matching
on the request method.

There is also the majority of the initial SiteBuilder work done. The
idea behind a site builder was to point it at a directory and have it
turn .tsx files into a SPA, .ts files into api routes, and everything
else into static assets. It's construction should be extensible enough
that new RouteBuilders can be created externally so that other filetypes
can be turned into routes. ie. Markdown files can be pre-rendered, image
files can be compressed, etc etc. It's also meant to be generic enough
that it can be run on any platform and not just on Deno.

Currently, the SiteBuilder only has a static route builder. The
following is intended to be done soon:

* Server route builders should be straightforward, but I'd like to add
  some degree of type checking to make them break at app start and not
  at runtime.
* Client route builders are simple enough to build, since they will all
  serve the same index file.. but there is a chicken and egg issue here.
* I need to spec types for a ClientBuilder/ClientBundler that takes the
  array of ClientRoutes and turns them into a bundle! Hopefully it's
  easy enough.

Once these features are complete and test coverage is at 100% I will
likey take a break from pick for awhile, but I do want to return at a
later date to implement the following features:

* Routes should take optional input/output schemas that match OpenAPI
  definitions. Or default to unknown.
* Routers should have a get_description method that returns a sorted
  list of all Routes without their Handlers.
* We should have some tooling for converting a Route list to a sitemap
  or an OpenAPI spec.
* SiteBuilder should have an option to mount an OpenAPI spec or a
  sitemap at unused paths (and probably write those files into the root
  site path).

All in all, a good amount of cleanup and work is done now. Should have
the bare minimum of what I need soon. I think capnweb is going to pull
me away from swagger/openapi generators and client generators for
awhile. This SiteBuilder approach resolves tons of issues with cors,
versioning, and type safety.